Initial commit of script for cyanogenmod small /datadata workaround
Synopsis:

On CM 7.1 private app data is stored on /datadata partition which is an MTD
formatted as a yaffs filesystem.  This partition is fast (low lag) but
small, about 170MB.  If apps download additional data files, or cache large
amounts of data (e.g. G+, Youtube) then this partition can fill up and apps
will start to force close.

The /data partition is a very large partition, about 2GB, but access to it
is much slower and can cause apps to start slow, or pause during their
operation, which is annoying.  The /data partition is where all apps are
actually installed, and the /data/data directory is actually a symbolic link
to /datadata.

So, a workaround is to move the contents of the /datadata partition into a
real directory at /data/data on the large 2GB partition, but then to move
the files that are accessed frequently by an app back onto the fast
/datadata partition.  The files to move back are primarily the sqlite
database files and xml config files.

This script automates the process of move the desired files/directories from
the /data/data/ directory to the /datadata directory and creating symbolic
links to point to them.  The initial step, of moving the contents of
/datadata to /data/data is not performed by this script, but can be done by
hand fairly easily.

Here is the post on the cyanogenmod forums announcing this script:

   http://forum.cyanogenmod.com/topic/39718-workaround-for-small-datadata-partition/